H2fun(bipartite)	R Documentation

Specialisation of a bipartite web.

Description

Calculates the overall level of specialisation of all partners in a bipartite web.

Usage


H2fun(web)

Arguments

web 	 A matrix representing the interactions observed between higher trophic level species (columns) and lower trophic level species (rows). Usually this will be number of pollinators on each species of plants or number of parasitoids on each species of prey.

Details

H2' is an index describing the level of specialisation, or should one say: selectiveness, of an entire bipartite network (Blüthgen et al. 2006). It describes to which extent observed interactions deviate from those that would be expected given the species marginal sums. The more selective a species, the larger is H2' for the web.

H2' is an extension of d' (see dfun) for the entire network. Check there for details on finding maximum and minimum H2 values for a given web.

For internet implementation see also: http://nils.mib.man.ac.uk/~nils/stat/

Value

H2 	 The H2'-value for the web matrix.

H2min 	 Heuristic minimum H2-value for the web matrix.

H2max 	 Heuristic maximum H2-value for the web matrix.

H2uncorr 	 Uncorrected H2-values (before ranging between min and max), rounded to three digits.

Author(s)

Carsten F. Dormann and Jochen Fründ

References

Blüthgen, N., Menzel, F. and Blüthgen, N. (2006) Measuring specialization in species interaction networks. BMC Ecology 6, 9

See Also

dfun doing the same for each species in the web matrix.

Examples


data(Safariland)
H2fun(Safariland)

[Package bipartite version 0.92 Index]