V.ratio(bipartite)	R Documentation

Calculates the variance-ratio as suggested by Schluter (1984)

Description

A of species association is provided by the ratio of the variance in total species number (or total density of individuals) in samples to the sum of the variances of the individual species.

Usage


V.ratio(web)

Arguments

web 	 A matrix with pollinators in columns and plants in rows. For biogeographical applications: rows are islands (or sites).

Details

This is a rather straight-forward index, which is described and evaluated extensively in Schluter (1984). He also warns against overinterpretation of the value. In principle, V-ratios larger than 1 indicate positive, smaller than 1 negative associations. Ecologically, competition can lead to small or large values, depending on their exact effects (see discussion in the Schluter paper).

Value

Returns the V-ratio, i.e.~a single value representing the ratio of variance in species number and variance in individual numbers within species.

Note

Any quantitative matrix is first transformed into a binary (presence-absence) matrix!

Do not interpret without first reading the paper! It's worth it! See also applications in other studies, such as Gotelli and Rohde (2002).

Author(s)

Carsten F. Dormann

References

Gotelli, N.J. and Rohde, K. (2002) Co-occurrence of ectoparasites of marine fishes: a null model analysis. Ecology Letters 5, 86–94

Schluter, D. (1984) A variance test for detecting species associations, with some example applications. Ecology 65, 998–1005

See Also

C.score for another measure of species associations.

Examples


data(Safariland)
V.ratio(Safariland)

[Package bipartite version 0.92 Index]