elberling1999(bipartite)	R Documentation

No. of visits in a pollination web of arctic-alpine Sweden.

Description

This study took place in the subarctic alpine zone of Latnjajaure, in northern Sweden. Field work was conducted from May 21 to August 23, 1994. The objective was to describe the plant-flower visitor interaction matrix of this area and compare it with the characteristics of other subarctic alpine systems and with pollination systems of lower latitudes, especially in relation to the role of flies as flower visitors at high latitudes.

The authors recorded their data by counting the number of visits of each flower visitor species to each plant species. Regardless of whether insects were observed to forage for nectar and pollen or to perform sun-basking, they were all classified as flower visitors and potential pollinators and the plant species visited were recorded. Data are presented as an interaction frequency matrix, in which cells with positive integers indicate the frequency of interaction between a pair of species, and cells with zeros indicate no interaction.

See also http://www.nceas.ucsb.edu/interactionweb/html/datasets.html

Usage

data(elberling1999)

Format

A data frame with 12 plant species (in rows) and 102 pollinators (columns).

References

Elberling H. and Olesen J.M. (1999) The structure of a high latitude plant-flower visitor system: the dominance of flies. Ecography 22, 314–323

Examples


data(barrett1987)
## maybe str(barrett1987) ; plot(barrett1987) ...

[Package bipartite version 0.92 Index]