extinction(bipartite)	R Documentation

Simulates extinction of a species from a bipartite network

Description

Following (how I remember) the paper of Memmott et al. (2004), this function deletes a column (e.g. pollinator) or row (e.g. plant). Only a helper function for second.extinct, really.

Usage


extinction(web, participant = "both", method = "random")

Arguments

web 	 A matrix representing the interactions observed between higher trophic level species (columns) and lower trophic level species (rows). Usually this will be number of pollinators on each species of plants or number of parasitoids on each species of prey.

participant 	 Which level of participant to remove: lower removes a row, higher removes a row, both randomly picks either row or column. Partial matching of strings implemented.

method 	 Determines sequence of extinctions: random removes a random participant, while abundance removes the least abundant species first. Partial matching of strings implemented.

Details

In itself rather useless. Called repeatedly by second.extinct to build an extinction sequence and accordingly a sequence of secondary extinctions.

Value

Returns the same matrix that was given as object, just with one row or column being turned into zeros.

Author(s)

Carsten F. Dormann

References

Memmott, J., Waser, N. M. and Price, M. V. 2004 Tolerance of pollination networks to species extinctions. Proceedings of the Royal Society B 271, 2605–2611

See Also

second.extinct

Examples


data(Safariland)
(w <- extinction(Safariland, participant="low", method="abun"))
empty(w, count=TRUE)

[Package bipartite version 0.92 Index]