genweb(bipartite)	R Documentation

Generate a random bipartite web

Description

Generates a random bipartite web, based on r2dtable and lognormal marginal distributions.

Usage


genweb(N1 = 10, N2 = 30, dens = 2)

Arguments

N1 	 Number of species in the lower trophic level; or a vector of length 2 giving number of lower and higher trophic level species. Defaults to 10.

N2 	 Number of species in the higher trophic level. Ignored if N1 a vector of length 2. Defaults to 30.

dens 	 Interaction density, i.e.~how many interactions there shall be, on average, for each species. Defaults to 2 (the median observed interaction density in the NCEAS pollination webs).

Details

This function can be used to create simple, but not necessarily realistic, bipartite webs for given dimensionality and interaction density. Marginal distributions are assumed to be lognormal, mean and standard deviation are calculated from N1, N2 and dens (see code for details).

Value

A matrix with N1 x N2 species.

Note

Can be a bit time-consuming for large webs, because the absolute values for both dimensions have to match perfectly. This involves a rather unelegant while-loop.

Author(s)

Jochen Fründ and Carsten F. Dormann

Examples


genweb()

[Package bipartite version 0.92 Index]