inouye1988(bipartite)	R Documentation

A pollination network.

Description

This data set reports a community-level study of the pollination biology of alpine plants in Kosciusko National Park in the Snowy Mountains of south-eastern New South Wales, Australia. The flora and their associated insect pollinators were observed from December 1983 until March 1984.

Usage

data(inouye1988)

Format

A data frame with 41 observations on the following 83 variables, with plant species in rows and pollinators in columns.

Details

The authors recorded their data by counting the number of individual flower visitors caught on each plant species. The total number of individuals collected on each plant species provide a rough estimate of the level of visitation that each species received. Data are presented as an interaction frequency matrix, in which cells with positive integers indicate the frequency of interaction between a pair of species, and cells with zeros indicate no interaction.

Note

Male and female pollinators were summed when moving this data set from NCEAS to bipartite.

Source

NCEAS data base on interaction webs: http://www.nceas.ucsb.edu/interactionweb/html/datasets.html

References

Inouye, D.W. and G.H. Pyke (1988) Pollination biology in the Snowy Mountains of Australia: comparisons with montane Colorado, USA. Australian Journal of Ecology 13: 191–210.

Examples


data(inouye1988)
plotweb(inouye1988)

[Package bipartite version 0.92 Index]