kato1990(bipartite)	R Documentation

No. of individuals caught in a pollination web of a Japanese beech forest.

Description

The study took place at the Kyoto University Forest of Ashu, at the northeastern boundry of the Kyoto Prefecture in Japan, between 1984 and 1987. The paper deals with the flowering phenology of 91 plant species, the community structure of flower-visiting insects, and the spectrum of floral hosts for flower visitors. The emphasis is laid on the pattern of community organization of flower-visiting insects in a primary forest ecosystem of western Japan.

The authors recorded their data by counting the number of individual flower visitors caught on each plant species. The total number of individuals collected on each plant species provide a rough estimate of the level of visitation that each species received. Data are presented as an interaction frequency matrix, in which cells with positive integers indicate the frequency of interaction between a pair of species, and cells with zeros indicate no interaction. For details and data see http://www.nceas.ucsb.edu/interactionweb/html/datasets.html

Usage

data(kato1990)

References

Kato, M., T. Makutani, T. Inoue, and T. Itino. 1990. Insect-flower relationship in the primary beech forest of Ashu, Kyoto: an overview of the flowering phenology and seasonal pattern of insect visits. Contr. Biol. Lab. Kyoto Univ. 27:309-375.

Examples


data(kato1990)
## maybe str(kato1990) ; plot(kato1990) ...

[Package bipartite version 0.92 Index]