kevan1970(bipartite)	R Documentation

A pollination network.

Description

The total number of individuals collected on each plant species provide a rough estimate of the level of visitation that each species received.

Usage

data(kevan1970)

Details

General information

This study sought to determine the importance of insect-flower relations to both plants and insects in a high arctic community as well as the degree to which some of the more common arctic plants are dependent on insects for pollination and reproduction. The research was conducted in 1967 at Hazen Camp (81 49'N, 71 18' W) near Lake Hazen on Northern Ellesmere Island, the most northerly island of the Canadian Arctic Archipelago.

Data type

The authors recorded their data by counting the number of individual flower visitors caught on each plant species. The total number of individuals collected on each plant species provide a rough estimate of the level of visitation that each species received. Data are presented as an interaction frequency matrix, in which cells with positive integers indicate the frequency of interaction between a pair of species, and cells with zeros indicate no interaction.

References

Kevan, P. G. 1970. High Arctic Insect-Flower Visitor Relations: The Inter-Relationships of Arthropods and Flowers at Lake Hazen, Ellesmere Island, Northwest Territories. University of Alberta, Canada.

Examples


data(kevan1970)

[Package bipartite version 0.92 Index]