memmott1999(bipartite)	R Documentation

Pollination network.

Description

This study was conducted in a 150 x 250 m medow plot in the vicinity of Bristol, U.K. in July 1997. The objective was to describe the plant-flower visitor interaction web of this area, taking into account species abundances and their frequency of interaction. Twenty five plant species were studied, and 79 flower visitor species were recorded visiting them.

Usage

data(memmott1999)

Details

The author recorded her data by counting the number of visits of each flower visitor species to each plant species, and by independently measuring the abundance of plant and animal taxa. Data are presented as an interaction frequency matrix, in which cells with positive integers indicate the frequency of interaction between a pair of species, and cells with zeros indicate no interaction.

Source

NCEAS

References

Memmott, J. 1999. The structure of a plant-pollinator food web. Ecology Letters 2:276-280.

Examples


data(memmott1999)
## maybe str(memmott1999) ; plot(memmott1999) ...

[Package bipartite version 0.92 Index]