mosquin1967(bipartite)	R Documentation

Pollination network.

Description

This study took place on Melville Island, N.W.T., Canada from July 19 to July 31 1965. While collecting plants the authors made some observations on the occurence and behavior of flower visiting insects as well as on the scent and other target characteristics of flowers.

Usage

data(mosquin1967)

Details

The authors recorded their data by counting the number of individual flower visitors caught on each plant species. The total number of individuals collected on each plant species provide a rough estimate of the level of visitation that each species received. Data are presented as an interaction frequency matrix, in which cells with positive integers indicate the frequency of interaction between a pair of species, and cells with zeros indicate no interaction.

Source

NCEAS

References

Mosquin, T., and J. E. H. Martin. 1967. Observations on the pollination biology of plants on Melville Island, N.W.T., Canada. Canadian Field Naturalist 81:201-205.

Examples


data(mosquin1967)
## maybe str(mosquin1967) ; plot(mosquin1967) ...

[Package bipartite version 0.92 Index]