nested(bipartite)	R Documentation

Calculates any of several measures of nestedness

Description

Wrapper function calling one, several or all currently implemented nestedness measures

Usage


nested(web, method = "binmatnest2", ..., rescale = FALSE)

Arguments

web 	 A matrix with elements of a set (e.g., plants) as rows, elements of a second set (e.g., pollinators) as columns and number of interactions as entries.

method 	 One or more of the following: discrepancy, discrepancy2, binmatnest, binmatnest2, NODF, NODF2, C.score, checker, wine, ALL. See details for details on each method

... 	 Arguments passed on to other nestedness functions. Options need to be specified (i.e. no positional parsing).

rescale 	 Should all measures be rescaled so that higher values mean higher nestedness? Defaults to FALSE, i.e. the standard interpretation of each measure is maintained.

Details

There are five different measures currently available:

binmatnest calculates nestedness temperature following the function nestedness (0 = cold = highly nested; 100 = hot = not nested at all). It uses the original program of Miguel Rodríguez-Gironés, only called from R; binmatnest2, in contrast, is the implementation in nestedtemp of the same algorithm by Jari Oksanen. Because binmatnest sometimes (and to us unexplicably) invert the matrix, we prefer the binmatnest2 option.

Discrepancy calculates the number of non-nested 0s and 1s. While discrepancy calls the function with the same name, discrepancy2 calls nesteddisc, which handles ties differently. Most of the time, these two should deliver very, very similar results. Higher values indicate lower nestedness.

NODF is the nestedness measure proposed by Almeida-Neto et al., correcting for matrix fill and matrix dimensions. Values of 0 indicate non-nestedness, those of 100 perfect nesting. NODF2 sorts the matrix before calculating the measure. NODF is, I understand, closer to the version presented in the paper, while NODF2 seems to make more sense for comparisons across different networks (because it is independent of the initial presentation of the matrix). Both call nestednodf in vegan. (Yes, I initially programmed NODF myself, only to find that it was there already. Luckily, there was a perfect agreement between my (depricated) version and nestednodf.)

C.score calculates the number of checkerboard pattern in the matrix. As default, it normalises this value between min and max, so that values of 0 indicate no checkerboards (i.e. nesting), while a value of 1 indicates a perfect checkerboard. checker is the non-normalised version, based on nestedchecker.

wine is the only nestedness measure using the information on the weight of a link. See wine for details.

Value

A vector with values for each of the selected nestedness measures.

Note

The idea behind this function is to encourage the comparison of different nestedness measures. That does not mean, we necessarily see much ecological sense in them (see, e.g., the paper by Blüthgen et al. 2008).

Author(s)

Carsten F. Dormann carsten.dormann@ufz.de

References

Almeida-Neto, M., Gumaraes, P., Gumaraes, P.R., Loyola, R.D. and Ulrich, W. 2008. A consistent metric for nestedness analysis in ecological systems: reconciling concept and measurement. Oikos 117, 1227–1239.

Blüthgen, N., J. Fründ, D. P. Vázquez, and F. Menzel. 2008. What do interaction network metrics tell us about specialisation and biological traits? Ecology 89, 3387–3399.

Brualdi, R.A. and Sanderson, J.G. 1999. Nested species subsets, gaps, and discrepancy. Oecologia 119, 256–264.

Galeano, J., Pastor, J.M., Iriondo and J.M. 2008. Weighted-Interaction Nestedness Estimator (WINE): A new estimator to calculate over frequency matrices. arXiv 0808.3397v2 [physics.bio-ph]

Rodriguez-Girones, M.A. and Santamaria, L. 2006. A new algorithm to calculate the nestedness temperature of presence-absence matrices. J. Biogeogr. 33, 924–935.

Stone, L. and Roberts, A. 1990. The checkerboard score and species distributions. Oecologia 85, 74–79.

See Also

C.score, wine, nestedness, discrepancy; and, within vegan: nestedtemp, nestedchecker, nesteddisc, nestednodf

Examples


data(Safariland)
nested(Safariland, "ALL")
nested(Safariland, "ALL", rescale=TRUE)
# illustration that non-normalised C.score and checker are the same:
nested(Safariland, c("C.score", "checker"), normalise=FALSE)

[Package bipartite version 1.01 Index]