networklevel(bipartite)	R Documentation

Analysis of bipartite webs at the level of the entire network

Description

Calculates a variety of indices and values for a bipartite network.

Usage


networklevel(web, index="ALL", ISAmethod="Bluethgen", 
    SAmethod = "Bluethgen", extinctmethod = "r", nrep = 100, 
    plot.it.extinction = FALSE, plot.it.dd=FALSE, CCfun=median, 
    dist="horn", normalise=TRUE, empty.web=TRUE, logbase="e", 
    intereven="sum")

Arguments

web 	 Web is a matrix representing the interactions observed between higher trophic level species (columns) and lower trophic level species (rows). Usually this will be number of pollinators on each species of plants or number of parasitoids on each species of prey.

index 	 One or more of the following (exact match only!): “number of species”, “connectance”, “web asymmetry”, “links per species”, “number of compartments”, “compartment diversity”, “cluster coefficient”, “degree distribution”, “mean number of shared partners”, “togetherness”, “C score”, “V ratio”, “discrepancy”, “nestedness”, “weighted nestedness”, “extinction slope”, “robustness”, “ISA”,“SA”, “niche overlap”, “generality”,“vulnerability”, “linkage density”, “Fisher alpha”, “mean interaction diversity”, “interaction evenness”, “Alatalo interaction evenness”, “diversity”, “H2”; “ALL” (default) calculates all indices (returning a list (since degree distribution fits are a table within this list and cannot be simplified)), “ALLBUTDD” calculates all indices except degree distribution fits. This latter has the advantage that the output can be simplified to a vector; “info” returns more general information on the network, “binary” returns a best-of selection of indices based on a binary network, “quantitative” returns a best-of selection of indices based on quantitative networks, “topology” returns indices more abstractly describing network properties. Also CHECK details below!

ISAmethod 	 Method to use for calculating interaction strength (= dependence) asymmetry; original by “Bascompte” is biased for singletons and few interactions (range 0 to infty); “Bluethgen” (default) excludes singletons and corrects for low number of interactions (range -1 to 1); see Bascompte et al. (2006) and Blüthgen et al. (2007).

SAmethod 	 Which method to use to calculate the specification asymmetry: mean of log-transformed dependencies (“log”) or Blüthgen's abundance-weighted mean (default); see Blüthgen et al. (2007).

extinctmethod 	 Specifies how species are removed from matrix: “random” or “abundance” (partial matching), where abundance removes species in the order of increasing abundance (i.e. rarest first); see Memmott et al. (2004).

nrep 	 Number of replicates for the extinction sequence analysis.

plot.it.extinction 	 logical; plot the extinction sequence graph; defaults to FALSE.

plot.it.dd 	 logical; plot the degree distribution fits?; defaults to FALSE.

CCfun 	 Method to use when calculating the clustering coefficient. Originally proposed as mean of cluster coefficients for each node. Defaults to “median”, because cluster coefficients are strongly skewed.

dist 	 Distance metric to be used to calculate niche overlap; defaults to Horn's index, which is the recommendation of Krebs (Ecological Methodology); for other options see vegdist in vegan.

normalise 	 Logical; shall the C-score and togetherness metrics be normalised to a range of 0 to 1? Defaults to TRUE.

empty.web 	 Shall the empty columns and rows be deleted? Defaults to TRUE.

logbase 	 Shall the various diversity indices be calculated to the base of e (default) or 2? Log2 is the proposal for generality and vulnerability by Bersier et al. (2002), while Shannon uses ln. The choice of the base will not affect the results qualitatively, only by a scaling factor.

intereven 	 Shall all cells of the matrix be used to calculate the interaction evenness (intereven="prod")? Or, as given by Bersier et al. (2002) and Tylianakis et al. (2007), should only the realised links be used (intereven="sum"; default)? Prod and sum refer to using the log of the product of matrix dimensions (i.e. all cells) or the log of the sum of non-zero cells (i.e. number of links) as denominator in the evenness formula. See last paragraph of the details-section for views on these two options!

Details

This function implements a variety of the many (and still procreating) indices describing network topography. Some are embaracingly simple and mere descriptors of a network's outer appearance (such as number of species in each trophic level or the number of links (= non-zero cells) in the web). Others are variations on Shannon's diversity index applied to within column or within rows. Only extinction slope is newly implemented here, and hence described in a bit more detail.

Note that Bersier et al. (2002) have three levels of values for some of their indices: qualitative (i.e. based on binary networks), quantitative (based on networks with information on the number of interactions observed for each link), and weighted-quantitative (where also independent information on the abundances of the participants is fed into the formula). At present, we implement only qualitative and quantitative indices! One exception is d', which can be called with an abundance vector (in function dfun). You can also get the qualitative version of quantitative indices (such as vulnerability) by simply calling networklevel on ``binarised'' data: networklevel(Safariland>0, index="vulnerability"). (Why you would want to do that, however, is currently beyond me.)(Thanks to Jason Tylianakis for proposing to put this clarification into the help!)

Extinction slope works on a repeated random sequence of species extinctions (within one trophic level), and calculates the number of secondary extinctions (in the other level). These values are then averaged (over the nrep runs) and plotted against the number of species exterminated. The proportion still recent (on the y-axis) regressed against the proportion exterminated (on the x-axis) is hence standardised to values between 0 and 1 each. Through this plot, a hyperbolic regression is fitted, and the slope of this regression line is returned as an index of extinction sensitivity. The larger the slope, the later the extinction takes its toll on the other trophic level, and hence the higher the redundancy in the trophic level under consideration. Using plot.it=F also returns the graphs (set history to recording in the plotting window). Changing the extinctionmethod to “abundance” will always result in the same sequence (by increasing abundance) and hence does not require replication.

Most indices are straightforward, one-line formulae; some, such as H2', also require a re-arranging of the matrix; and one, secondary extinction slope, internally requires iterative runs, making the function relatively slow. If you are not interested in the secondary extinction slopes, simply set nrep=1 to make it much faster.

Sum or Prod: How to calculate interaction evenness? I shall first put down my argument for ``prod'' and then Jason Tylianakis' arguments for ``sum''. Carsten: ``I do not want to defend a position I cannot hold against the flood of qualified criticism, and shall be happy to change the default to option ``sum'' (i.e. Jason's proposal). Nevertheless, I shall make a very brief attempt to defend my (and Nico's point of view). Imagine a completely different situation: I have ``counted'' birds in a landscape. From a more meticulous colleague I know that there are 27 bird species breeding at the moment, but on that two morning that I went out, I could only hear 15. Now I want to calculate the Shannon diversity (and evenness) of birds in that landscape. The ``normal'' (in the sense of established) approach to use the data from my 15 species. But hold on: I KNOW there are more species out there. I don't know how many (i.e. there may be more than the 27 my colleague has found), but there are at least 27. If I only use the data from my 15 species, I will get a higher evenness value than when I also include the 12 zeros. My conclusion would be: I don't want to overestimate evenness only because I couldn't look long enough, thus I use all 27 values.'' Jason: ``I would disagree because what you ``know'' is based on your meticulous colleague's `sampling', which will also have its limits. If all you wanted was to know the total number of species there (assuming none have gone extinct), then what you propose is fine. However, the problem comes when you want to compare sites, and then sampling effort should be standardised. In most cases we know we don't have a full representation of the diversity (or food web) of an area, but we know for a given spatial or temporal sampling scale that one site differs from another in certain ways, and to me that is the most important. Anyway, it is all a question of scale and the precise question being asked. So what about making it an option in bipartite that you can either choose to divide by the realised links (give our 2007 paper as a ref, so people know it's comparable to that) or divide by the number of potential links, if that's the question people want to ask?'' There you go: it's your choice!

Value

Depending on the selected indices, some or all of the below (returned as vector is “degree distribution” was not requested, otherwise as list):

number of higher trophic species 	 Number of species in the higher trophic level, i.e. ncol(web).

number of lower trophic species 	 Number of species in the lower trophic level, i.e. nrow(web).

connectance 	 Realised proportion of possible links (Dunne et al. 2002): sum of links divided by number of cells in the matrix (= number of higher times number of lower trophic level species). This is the standardised number of species combinations often used in co-occurrence analyses (Gotelli & Graves 1996)

web asymmetry 	 Balance between numbers in the two levels: positive numbers indicate more lower-trophic level species, negative more higher-trophic level species; implemented as (nrow(web)-ncol(web))/sum(dim(web)); web asymmetry is a null-model for what one might expect in dependence asymmetry: see Blüthgen et al. (2007).

links per species 	 Mean number of links per species (qualitative): sum of links divided by number of species.

number of compartments 	 Compartments are sub-sets of the web which are not connected (through either higher or lower trophic level) to another compartment. Mathematically, they are Jordan blocks, but this implementation is rule-based (and fast). They are also nicely visualised in the visweb function.

compartment diversity C.D. 	 Shannon's diversity index across compartment sizes (i.e. number of participants); see Tylianakis et al. (2007).

cluster coefficient 	 The CC for a network is the average CC of its members. CC for each node, in turn, is simply the number of realised links devided by the number of possible links. Introduced by Watts & Strogatz (1998) and described in Wikipedia under http://en.wikipedia.org/w/index.php?title=Clustering_coefficient.

degree distribution 	 Coefficients and fits for three different functions to degree distributions: exponential, power law and truncated power law. See degreedistr for details and references.

mean number of shared hosts 	 The simplest measure of co-occurrence and hence similarity in host preferences of the higher trophic level; based on Roberts & Stone (1990) and Stone & Roberts (1992).

mean number of shared predators 	 Same as before, only for the lower trophic level.

togetherness 	 Mean number of co-occupancies across all species-host-combinations; the whole matrix is scanned for submatrices of the form c(0,0,1,1), representing perfect matches of co-presences and co-absences. These are counted for each pairwise species combination, and averaged. Based on Stone & Roberts (1992).

C score 	 Mean (normalised) number of checkerboard combinations across all higher trophic level species. Values close to 1 indicate that there is evidence for disaggregation, e.g.~through competition. Value close to 0 indicate aggregation of species (i.e.~no repelling forces between species). See Stone and Roberts (1990) for details.

V ratio 	 Variance-ratio of species numbers to individual numbers within species for the higher trophic level. Values larger than 1 indicate positive aggregation, values between 0 and 1 indicate disaggregation of species. See Schluter (1984) for details.

discrepancy 	 Discrepancy as proposed by Brualdi & Sanderson (1999); see [discrepancy for details.

nestedness 	 Nestedness temperature of the matrix (0 means cold, i.e.~high nestedness, 100 means hot, i.e.~chaos). For details see nestedness and Rodríguez-Gironés & Santamaria (2002). Notice that the function nestedness does not calculate any null model, simply because it is too computer-intensive. networklevel calls nestedtemp! If you are interested in the different null models, please use the function nestedness or nestedtemp in vegan directly.

weighted nestedness 	 A nestedness version that considers interaction frequencies (and is hence weighted), proposed by Galeano et al. (2007) and implemented in wine. It ranges between 1 (perfect nestedness) and 0 (perfect chaos). Note that this is the OPPOSITE interpretation of nestedness temperature!

dependence asymmetry/interaction strength asymmetry 	 Explaining dependence asymmetry will require more space than just a few lines. In essence, it is also a measure of specialisation, across both trophic levels. Proposed by Bascompte et al. (2006) and critiqued and alterations proposed by Blüthgen et al. (2007). The latter also show that dependence asymmetry can be almost entirely explained by web asymmetry (see above). Positive values (only possible of DAmethod="Bluethgen") indicate higher dependence in the higher trophic level.

specialisation asymmetry 	 Asymmetry (higher vs. lower trophic level) of specialisation now based on d' (see dfun), which is insensitive to the dimensions of the web. Again, two options of calculation are available: the one proposed by Blüthgen et al. (2007), where they weight the specialisation value for each species by its abundance (SAmethod="Bluethgen") or where d'-values are log-transformed (argueing that d'-values are indeed log-normally distributed: SAmethod="log"). Since the mean d-value for the lower trophic level is subtracted from that of the higher, positive values indicate a higher specialisation of the higher trophic level.

extinction slope higher trophic level 	 Slope of the secondary extinction sequence in the higher trophic level, following extermination of species in the lower trophic level; see details.

extinction slope lower trophic level 	 Slope of the secondary extinction sequence in the lower trophic level, following extermination of species in the higher trophic level; see details.

robustness 	 Calculates the area below the extinction curve; see robustness for details.

niche overlap 	 Mean similarity in interaction pattern between species of the same trophic level, calculated by default as Horn's index. Values near 0 indicate no common use of niches, 1 indicates perfect niche overlap. (In how far it makes sense for hosts of predators to call their commonness in enemies “niche overlap” is a different issue. There are people calling predators “negative resources” (couldn't be asked to look up the reference). I would probably rather call it similarity in predators.)

generality 	 Effective mean number of prey per predator; see Tylianakis et al. (2007).

vulnerability 	 Effective mean number of predator per prey; see Tylianakis et al. (2007).

linkage density 	 Mean number of interactions per species (quantitative); see Tylianakis et al. (2007).

Fisher's alpha 	 An alternative measure of interaction diversity (using fisherfit from vegan).

HTL mean interaction diversity 	 Mean diversity of prey per predator. HTL refers to Higher Trophic Level. HTLmid is analog to Shannon's diversity index for interactions (see below), but calculated for each species in the HTL, then averaged.

LTL mean interaction diversity 	 Mean diversity of predators per prey. LTL refers to Lower Trophic Level. LTLmid is analog to Shannon's diversity index for interactions (see below), but calculated for each species in the LTL, then averaged.

interaction evenness 	 Shannon's evenness for the web entries, treating zeros as no data. Note that the two options are rather different. Here a brief attempt to justify our default "prod": By definition, IE = H/Hmax; H = -sum(p.i.mat*log(p.i.mat)), where p.i.mat = matrix/sum(entries in matrix). This means, when calculating H, we treat the matrix cells (=links) as species, and the interactions (cell values) as measure of their abundance. By definition, Hmax = ln(N). The key question is: What is the right value for N? Since we treat the matrix cells as species, it is (clearly?) the number of matrix cells, i.e. number of higher trophic level species x number of lower trophic level species. What else? Were we to use the interpretation of Tylianakis et al. (2007), then Hmax = ln(sum of entries in matrix). This means, we equate ``number of interactions'' (another phrase for sum of matrix entries) = ``number of species''. That means, each interaction is a species. What should that mean? Why should that measure ``interactions evenness''? Why should we move from a view of ``cells are species'' when calculating H to a view of ``interactions are species'' when calculating N? To say the least, it doesn't seem consistent.

Alatalo interaction evenness 	 A different measure for web entry evenness, as proposed by Müller et al. (1999).

Shannon diversity 	 Shannon's diversity of interactions (i.e. network entries).

H2 	 H2' is a network-level measure of specialisation. It ranges between 0 (no specialisation) and 1 (complete specialisation). H2' is a measure of discrimination, i.e. calculated in comparison of no specialisation (see H2fun for details. To avoid confusion of keys (apostrophe vs. accent), we call the H2' only H2 here.

Note

Some indices require rather long computation times on large webs. Particularly slow (for large webs) are: “extinction slope” and “nestedness”. For ordinary-sized webs (tens by tens), networklevel is reasonably fast. An example of a large web is kato1990: try it, possibly using system.time as a wrapper, and prepare to do some reading on the side ...

If you want to increase the speed by omitting some indices, here a rough guide: Ask only for the indices you are interested in! Otherwise, here is the sequence of most time-consuming indices: 1. The slowest function is related to extinction slopes and robustness. Excluding both makes the function faster. 2. Degree distributions are somewhat time consuming. 3. Fisher's alpha is computed iteratively and hence time consuming. 3. Nestedness and weighted nestedness are not the fastest of routines. 4. Number (and diversity) of compartments calls a recursive and hence relatively slow algorithm. 5. H2 and specialisation asymmetry require an iterative, heuristic search algorithm. Finally, excluding discrepancy can also moderately decrease computation time.

Author(s)

Carsten F. Dormann carsten.dormann@ufz.de

References

Bascompte, J., Jordano, P. and Olesen, J. M. 2006 Asymmetric coevolutionary networks facilitate biodiversity maintenance. Science 312, 431–433

Bersier, L. F., Banasek-Richter, C. and Cattin, M. F. (2002) Quantitative descriptors of food-web matrices. Ecology 83, 2394–2407

Blüthgen, N., Menzel, F., Hovestadt, T., Fiala, B. and Blüthgen N. 2007 Specialization, constraints and conflicting interests in mutualistic networks. Current Biology 17, 1–6

Burgos, E., H. Ceva, R.P.J. Perazzo, M. Devoto, D. Medan, M. Zimmermann, and A. Maria Delbue (2007) Why nestedness in mutualistic networks? Journal of Theoretical Biology 249, 307–313

Corso G, de Araújo AIL, de Almeida AM (2008) A new nestedness estimator in community networks. arXiv 0803.0007v1 [physics.bio-ph]

Dormann, C.F., Fründ, J., Blüthgen, N., and Gruber, B. (2009) Indices, graphs and null models: analysing bipartite ecological networks. The Open Ecology Journal 2, 7–24.

Dunne, J. A., R. J. Williams, and N. D. Martinez. 2002 Food-web structure and network theory: the role of connectance and size. Proceedings of the National Academy of Science USA 99, 12917–12922

Galeano J, Pastor JM, Iriondo JM (2008) Weighted-Interaction Nestedness Estimator (WINE): A new estimator to calculate over frequency matrices. arXiv 0808.3397v1 [physics.bio-ph]

Gotelli, N. J., and G. R. Graves. 1996 Null Models in Ecology. Smithsonian Institution Press, Washington D.C.

Memmott, J., Waser, N. M. and Price M. V. 2004 Tolerance of pollination networks to species extinctions. Proceedings of the Royal Society B 271, 2605–2611

Müller, C. B., Adriaanse, I. C. T., Belshaw, R. and Godfray, H. C. J. 1999 The structure of an aphid-parasitoid community. Journal of Animal Ecology 68, 346–370

Roberts, A. and Stone, L. 1990 Island-sharing by archipelago species. Oecologia 83, 560–567

Rodríguez-Gironés M.A., and Santamaría L. 2006. A new algorithm to calculate the nestedness temperature of presence-absence matrices. Journal of Biogeography 33, 924–935

Schluter, D. 1984 A variance test for detecting species associations, with some example applications. Ecology 65, 998-1005.

Stone, L. and Roberts, A. 1990 The checkerboard score and species distributions. Oecologia 85, 74–79.

Stone, L. and Roberts, A. 1992 Competitive exclusion, or species aggregation? An aid in deciding. Oecologia 91, 419–424

Tylianakis, J. M., Tscharntke, T. and Lewis, O. T. 2007 Habitat modification alters the structure of tropical host-parasitoid food webs. Nature 445, 202–205

Watts, D. J. and Strogatz, S. 1998 Collective dynamics of ‘small-world’ networks. Nature 393, 440–442

See Also

Some functions are implemented separately: H2fun, second.extinct, degreedistr, C.score and V.ratio

Examples


data(Safariland)
networklevel(Safariland)
networklevel(Safariland, index="ALLBUTDD") #excludes degree distribution fits

[Package bipartite version 1.03 Index]