olesen2002flores(bipartite)	R Documentation

A pollination network.

Description

The structure of pollination networks were investigated for two oceanic islands, the Azorean Flores and the Mauritian Ile aux Aigrettes.

Usage

data(olesen2002flores)

Details

The authors recorded their data by counting the number of visits of each flower visitor species to each plant species. Data are presented as an interaction frequency matrix, in which cells with positive integers indicate the frequency of interaction between a pair of species, and cells with zeros indicate no interaction.

References

Olesen, J. M., L. I. Eskildsen, and S. Venkatasamy. 2002. Invasion of pollination networks on oceanic islands: importance of invader complexes and endemic super generalists. Diversity and Distributions 8:181-192.

Examples


data(olesen2002flores)
## maybe str(olesen2002flores) ; plot(olesen2002flores) ...

[Package bipartite version 0.92 Index]