plotweb2(bipartite)	R Documentation

Visualize a tripartite interaction matrix (e.g. a tritrophic foodweb)

Description

Two two dimensional matrix are plotted as a tripartite graph.

Usage


plotweb2(web, web2, method = "cca", empty = FALSE, labsize = 1, ybig = 1,
    y_width = 0.1, spacing = 0.05, arrow="no", col.interaction="grey80",
    col.pred = "grey10", col.prey="grey10", lab.space=1, lablength = NULL, 
    sequence=NULL, low.abun=NULL,low.abun.col="green", high.abun=NULL, 
    high.abun.col="red", method2 = "cca", empty2 = TRUE, spacing2 = 0.05,  
    arrow2="no", col.interaction2="grey80", col.pred2 = "grey30", 
    col.prey2="grey20", lablength2 = NULL, sequence.pred2=NULL,low.abun2=NULL,
    low.abun.col2="green", high.abun2=NULL, high.abun.col2="red")

Arguments

web 	 Web is a matrix representing the interactions observed between higher trophic level species (columns) and lower trophic level species (rows). Usually this will be number of pollinators on each species of plants or number of parasitoids on each species of prey.

web2 	 The other web to be included.

method 	 Default method is cca, which leads to as few crossings of interactions as possible. The other option is normal, which leaves order as given by the matrix.

empty 	 logical; should empty columns or empty rows be omitted from plotting; defaults to true

labsize 	 factor for size of labels, default is 1

ybig 	 vertical distance between upper and lower boxes, default is 1

y_width 	 width of upper and lower boxes, default is 0.1

spacing 	 horizonatal distance between upper and lower boxes, default is 0.05

arrow 	 display type of connection between upper and lower boxes, options are up, down, both and no, default is no, which is a polygonal connection between boxes.

col.interaction 	 color of interaction, default is grey80.

col.pred 	 color of upper boxes, default is grey10.

col.prey 	 color of lower boxes, default is grey10.

lab.space 	 sometimes it is neccessary to add additional space for labels below und above of the boxes, so all labes are shown, default is 1.

lablength 	 number of characters of labels that should be plotted. If zero no labels are shown, default is NULL which plots the complete labels.

sequence 	 list of two with two names vectors: seq.pred and seq.prey, which specify the order in which species are plotted. Cannot be set for method="cca". Defaults to NULL, where the sequence remains as given or is determined by the CCA internally.

low.abun 	 Vector with independent abundance estimates for the lower trophic level, NULL if none exists.

low.abun.col 	 Colour for depicting the abundance estimates for the lower trophic level; defaults to green.

high.abun 	 Vector with independent abundance estimates for the higher trophic level, NULL if none exists.

high.abun.col 	 Colour for depicting the abundance estimates for the lower trophic level; defaults to red.

method2 	 Default method is cca, which leads to as few crossings of interactions as possible. The other option is normal, which leaves order as given by the matrix.

empty2 	 logical; should empty columns or empty rows be omitted from plotting; defaults to true

spacing2 	 horizonatal distance between upper and lower boxes, default is 0.05

arrow2 	 display type of connection between upper and lower boxes, options are up, down, both and no, default is no, which is a polygonal connection between boxes.

col.interaction2 	 color of interaction, default is grey80.

col.pred2 	 color of upper boxes, default is grey10.

col.prey2 	 color of lower boxes, default is grey10.

lablength2 	 number of characters of labels that should be plotted. If zero no labels are shown, default is NULL which plots the complete labels.

sequence.pred2 	 list of two with two names vectors: seq.pred and seq.prey, which specify the order in which species are plotted. Cannot be set for method="cca". Defaults to NULL, where the sequence remains as given or is determined by the CCA internally.

low.abun2 	 Vector with independent abundance estimates for the lower trophic level, NULL if none exists.

low.abun.col2 	 Colour for depicting the abundance estimates for the lower trophic level; defaults to green.

high.abun2 	 Vector with independent abundance estimates for the higher trophic level, NULL if none exists.

high.abun.col2 	 Colour for depicting the abundance estimates for the lower trophic level; defaults to red.

Value

Returns a window with a tripartite graph of a foodweb.

Author(s)

Bernd Gruber bernd.gruber@ufz.de

See Also

For a different plot of foodwebs see visweb and plotweb

[Package bipartite version 0.92 Index]