robustness(bipartite)	R Documentation

Robustness to species extinctions

Description

Calculates the area below the extinction curve generated by second.extinct.

Usage


robustness(object)

Arguments

object 	 An object of type class bipartite, usually generated by second.extinct.

Details

This function calculates the area below the extinction curve generated by second.extinct as a measure of the robustness of the system to the loss of species.

The curve, first proposed by Memmott et al. (2004), is based on the fact that if a given fraction of species of one guild (for instance, the pollinators) are eliminated, a number of species of the other guild (e.g. plants) which depend on their interactions become extinct. The slope and general shape of the curve provided a straightforward graphic description of the tolerance of a system to the extinction of its component species.

An improvement of Memmott et al.'s curve was developed by Burgos et al. (2007) by introducing a quantitative measure of robustness with a single parameter R, defined as the area under the extinction curve. It is intuitive that R = 1 corresponds to a curve that decreases very mildly until the point at which almost all animal species are eliminated. This is consistent with a very robust system in which, for instance, most of the plant species survive even if a large fraction of the animal species is eliminated. Conversely R = 0 corresponds to an ATC that decreases abruptly as soon as any species is lost. This is consistent with a fragile system in which, for instance, even if a very small fraction of the animal species is eliminated, most of the plants loose all their interactions and go extinct.

Value

Returns the robustness of the web to the removal of species.

Note

This index complements the information given by slope.bipartite, although it has the advantage of not being constrained by the shape of the particular curve (concave or convex).

Author(s)

Mariano Devoto mdevoto@agro.uba.ar

References

Burgos, E., H. Ceva, R.P.J. Perazzo, M. Devoto, D. Medan, M. Zimmermann, and A. Maria Delbue (2007) Why nestedness in mutualistic networks? Journal of Theoretical Biology 249, 307–313

Memmott, J., Waser, N. M. and Price, M. V. 2004 Tolerance of pollination networks to species extinctions. Proceedings of the Royal Society B 271, 2605–2611

See Also

second.extinct for generating the required input object and slope.bipartite for an alternative, but inferior measure

Examples


data(Safariland)
ex <- second.extinct(Safariland, participant="low", method="ran", nrep=100, details=FALSE)
robustness(ex)

[Package bipartite version 0.92 Index]