schemske1978(bipartite)	R Documentation

A pollination network.

Description

Populations of seven early flowering, low-growing, perennial woodland herbs were studied in a 24 hectare sit in Brownfield Woods, close to Urbana, Illinois, U.S.A. The sequence and variety of flowers visited by individual insects were recorded for the first foraging insect observed in a randomly selected, 1 m2 quadrat located within a 10x10 m grid that included representatives of several plant species.

Usage

data(schemske1978)

Details

The authors recorded their data by counting the number of visits of each flower visitor species to each plant species. Data are presented as an interaction frequency matrix, in which cells with positive integers indicate the frequency of interaction between a pair of species, and cells with zeros indicate no interaction.

References

Schemske, D. W., M. F. Willson, M. N. Melampy, L. J. Miller, L. Verner, K. M. Schemske, and L. B. Best. 1978. Flowering Ecology of Some Spring Woodland Herbs. Ecology 59:351-366.

Examples


data(schemske1978)
## maybe str(schemske1978) ; plot(schemske1978) ...

[Package bipartite version 0.92 Index]