shuffle.web(bipartite)	R Documentation

Shuffle web entries

Description

Shuffles (= relocates) entries in a web matrix whilst maintaining the dimensionality.

Usage


shuffle.web(web, N)

Arguments

web 	 An interaction matrix.

N 	 Number of desired shuffled matrices.

Details

This function is designed to behave similar to r2dtable, i.e. it returns a list of randomised matrices. In contrast to r2dtable is does not keep marginal sums constant!

This function is thought of as a nullmodel for the analysis of bipartite webs. It keeps two web properties constant: The number of interactions and the number of links (and hence connectance). A comparison of shuffle.web- and r2dtable-based webs allows to elucidate the effect of marginal sums.

Value

A list of N randomised matrices with the same dimensions as the initial web.

Note

shuffle.web is not an extremely intelligent nullmodel. You may want to think of a better one for your specific application!

Author(s)

Carsten F. Dormann <carsten.dormann@ufz.de>

References

This nullmodel can be thought of as a quantitative version of Fortuna & Bascompte (2006) “null model 1”:

Fortuna, M. A., and J. Bascompte. 2006. Habitat loss and the structure of plant-animal mutualistic networks. Ecology Letters 9: 281-286.

For a very nice and thorough overview of nullmodels in general see:

Gotelli, N. J., and G. R. Graves. 1996. Null Models in Ecology. Smithsonian Institution Press, Washington D.C.

For nullmodels and their application to webs/networks see, e.g.:

Vázquez, D. P., and M. A. Aizen. 2003. Null model analyses of specialization in plant-pollinator interactions. Ecology 84: 2493-2501.

Vázquez, D. P., C. J. Melián, N. M. Williams, N. Blüthgen, B. R. Krasnov, and R. Poulin. 2007. Species abundance and asymmetric interaction strength in ecological networks. Oikos 116: 1120-1127.

See Also

r2dtable

Examples



data(Safariland)

shuffle.web(Safariland, N=2)


[Package bipartite version 0.92 Index]