slope.bipartite(bipartite)	R Documentation

Slope of extinction simulation

Description

Fits a hyperbolic function to the extinction simulation of second.extinct.

Usage


slope.bipartite(object, plot.it = TRUE, ...)

Arguments

object 	 An object of class “bipartite”, usually generated by second.extinct.

plot.it 	 Logical; want to see the graph?

... 	 Graphical parameters passed on to the plot command used for plotting, NOT to the curve command used for overlaying the curve.

Details

Function scales extinction sequences to values between 0 and 1 for each participant. The x-axis of the graph features the proportion of exterminated participants, while the y-axis depicts the proportion of secondary extinctions. Since these curves usually follow a hyperbolic function (see examples in Memmott et al. 2004), this is fitted to the data.

At present, only a function of type y sim 1 - x^a is fitted (using nls), i.e. without offset. While usually this function provides very good fits, do check the graph and judge for yourself. Fitting this simple function makes its parameter ‘a’ a measure of extinction vulnerability. The more gradual the secondary extinctions, the lower the absolute value of ‘a’. Or, phrased differently, large absolute values of ‘a’ indicate a very abrupt die-off, indicative of high initial redundancy in the network.

Value

Returns one number, the exponent of the fitted hyperbolic model.

Note

This function is not as vigorously tested as it should probably be. It worked fine for large networks, but small ones may behave strangely, I fathom.

Note also that it CANNOT fit concave curves! The most extreme is a linear decline of secondary extinctions with primary extinctions. The function can be adapted to handle this, but only at the expense of fitting another parameter.

Author(s)

Carsten F. Dormann

References

Memmott, J., Waser, N. M. and Price, M. V. (2004) Tolerance of pollination networks to species extinctions. Proceedings of the Royal Society B 271, 2605–2611

See Also

second.extinct for generating the required input object.

Examples


data(Safariland)
ex <- second.extinct(Safariland, participant="low", method="ran", nrep=100,
details=FALSE)
slope.bipartite(ex)

[Package bipartite version 0.92 Index]