small1976(bipartite)	R Documentation

A pollination network.

Description

The study took place in the Mer Bleue peat bog of Ottawa, Canada in 1973. The paper is a preliminary evaluation of the pollination relationships of the major entomophilous plant species of the Mer Bleue.

Usage

data(small1976)

Details

The authors recorded their data by counting the number of individual flower visitors caught on each plant species. The total number of individuals collected on each plant species provide a rough estimate of the level of visitation that each species received. Data are presented as an interaction frequency matrix, in which cells with positive integers indicate the frequency of interaction between a pair of species, and cells with zeros indicate no interaction.

Source

http://www.nceas.ucsb.edu/interactionweb/html/small_1976.html

References

Small, E. 1976. Insect pollinators of the Mer Bleue peat bog of Ottawa. Canadian Field Naturalist 90:22-28.

Examples


data(small1976)
## maybe str(small1976) ; plot(small1976) ...

[Package bipartite version 0.92 Index]