togetherness(bipartite)	R Documentation

Calculates the number of identical co-presences and co-absences for species-on-islands

Description

Togetherness, or T-score, describes the level of similarity in the distributional pattern of two species. Originally proposed by Stone & Roberts (1992) for biogeographical situations can it also be applied e.g.~to pollinators on different host plants.

Usage


togetherness(web, normalise=TRUE, FUN = mean, ...)

Arguments

web 	 A matrix with binary or counted interactions/links, where the higher trophic level is represented by columns.

normalise 	 Logical; shall index be normalised to a range of 0-1?

FUN 	 The function to summarise species-pair T-scores with; defaults to mean.

... 	 Arguments passed on to FUN, especially na.rm=T or colours for hist.

Value

Returns the average (default) togetherness of all species combinations.

Author(s)

Carsten F. Dormann

References

Stone, L. and Roberts, A. (1992) Competitive exclusion, or species aggregation? An aid in deciding. Oecologia 91, 419–424

See Also

C.score for another of Stone & Roberts' indices.

Examples


(m <- matrix(c(0,1,0,0,1,1,0,1,1,0), ncol=2, byrow=TRUE))
togetherness(m)
# or, with two togethernesses:
(n <- matrix(c(0,1,1,0,1,1,0,0,1,1,0,1,0,1), ncol=2, byrow=TRUE))
togetherness(n, normalise=FALSE)

data(Safariland)
togetherness(m)

[Package bipartite version 0.92 Index]